// src/models/Types.tsx
export type MansType = {
  man_id: string;
  man_name: string;
  is_car: string;
  is_spec: string;
  is_moto: string;
};
export type AugmentedMansModelsType = {
  man_id: string;
  models: (
    | { model_group: ModelsType["model_group"]; sub_models: ModelsType[] }
    | ModelsType
  )[];
};
export type CatsType = {
  category_id: number;
  category_type: number;
  has_icon: number;
  title: string;
  seo_title: string;
  vehicle_types: number[];
};

export type ModelsType = {
  model_id: number;
  man_id: number;
  model_name: string;
  model_group: string;
  sort_order: number;
  cat_man_id: number;
  cat_model_id: number;
  cat_modif_id: number;
  is_car: boolean;
  is_moto: boolean;
  is_spec: boolean;
  show_in_salons: number;
  shown_in_slider: number;
};